Simulation_project_v3/requirements.txt
# 核心数值计算库
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0

# JIT加速（必需）
numba>=0.59.0

# 统计建模
copulas>=0.12.0
statsmodels>=0.14.0

# 优化算法
deap>=1.4.0

# 并行计算后端（支持闭包序列化）
loky>=3.4.0

# 配置文件
pyyaml>=6.0

# 可视化
matplotlib>=3.8.0
seaborn>=0.13.0

# 测试
pytest>=8.0.0
pytest-cov>=5.0.0

# 代码质量
black>=24.0.0
flake8>=7.0.0

